backend/requirements.txt
aiogram==3.13.1
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.2
openai==1.30.0
aiohttp==3.9.5
pytz==2024.1
aiosqlite==0.19.0
